04_Kubeflow_Pipeline_CICD_Examples/CICD_Pipeline_with_AI_Platform/pipeline/trainer_image/requirements.txt
fire
cloudml-hypertune
gcsfs
pandas==0.24.0
google-api-python-client==1.7.8
joblib==0.13.0
scikit-learn==0.20.2
